Move test code depending on Spring to separate module
It is difficult to depend on Spring 6 in test code in our published modules because we maintain a Java 8 baseline while Spring 6 requires Java 17. Moving this code to a separate module avoids that issue. There is still one usage of Spring in our normal test code for ModifiedClassPathClassLoader.
