+++ samples/micrometer-samples-spring-framework6/src/test/java/io/micrometer/samples/spring6/aop/HighCardinalityAnnotationHandler.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.observation.aop;
+package io.micrometer.samples.spring6.aop;
 
 import io.micrometer.common.KeyValue;
 import io.micrometer.common.annotation.AnnotationHandler;
